MISC-29 Hacky hack fix to work around Google's behaviour
On an AMP compatible device, when the user clicks a link in Google's search results, they'll append {{&ampcf=1}} to the URL (presumably because they expect it to go via their ping page instead of direct).

This makes sure that ends up in the querystring rather than the url path.

Longer term, might be better to rewrite the ping url rather than just overriding href though