Recheck links after "more" button press
Previously, if the user clicked the "more" button to load additional search results, the new results would still be AMP'd. To fix this, we need to recheck all links after the additional results have been added to the page.

Clicking "more" results in many fires of the `DOMNodeInserted` event. This commit re-runs the AMP check after such an event, with a 500ms debounce.