FKAMP-3 Add script to redirect certain sites/domains to non-AMP enforcing alternatives
This is primarily for things like Google News where it's non-trivial to un-AMP them.

This current version redirects Google News to Bing News (which isn't something I ever thought I'd find myself willingly doing)

The intention is that this will only ever be used for domains that use AMP but don't declare it, or don't use {{canonical}} properly. Google are currently the only ones I've seen doing that so far